zhejiang_market_data_crawler/requirements.txt
requests
pandas
pyyaml
sqlalchemy
pymysql
-e ..